+++ HRMapp/HRMapp.Models/ProductionWorker.cs
@@ -15,7 +15,7 @@ namespace HRMapp.Models
             Skillsets = new List<Skillset>();
         }
 
-        public ProductionWorker(int id, string firstName, string lastName, int phoneNumber, string emailAddress, string street, string houseNumber, string zipCode, string city, List<Skillset> skillsets, TeamLeader teamLeader)
+        public ProductionWorker(int id, string firstName, string lastName, Int64 phoneNumber, string emailAddress, string street, string houseNumber, string zipCode, string city, List<Skillset> skillsets, TeamLeader teamLeader)
             : base(id, firstName, lastName, phoneNumber, emailAddress, street, houseNumber, zipCode, city)
         {
             Skillsets = skillsets;
